Initial work on dynamic de jure title system
Currently affects:
Kingdom of Arixo
Kingdom of Dixieland
Empire of Mexico
Kingdom of Deitscherei
Kingdom of Hudsonia
Kingdom of the Maritimes

There's probably some idiotic typo in the event code, so be on the lookout for that. Also, I should go back and check if any of these titles are used for anything else, so events and decisions don't decide to randomly break.